feat: access granted setup roles
Airport Authority Personnel: These will be the users that have a very high access level. They will be able to manage not only Airlines and Flights, but Airports too!

Fleet Manager: This will be given to the Fleet managers. They maintain airplanes.

Travel Agent: This will be given to the travel agents who will use the system to book tickets on behalf of their passengers. They will also be able to manage passengers that they create.

Flight Crew Member: This role will be given to the users

Airplane Audits